added centercrop code
You can upload this file to google colaboratory where you can upload your own images to a directory. Copy and paste the directories file path to the section of the code and run the code so it can automatically center crop all the images and download them to another directory(where you specified its path). I haven't tried making a statedic for this model because I think you can use it without it.